siirrä parsinta omaan tiedostoon ja lisää tuli html formaatille
